🗺️ Atlas: [architectural change] Split Redb Cold Storage Blob
🕸️ Tangle: `src/storage/redb_cold_storage.rs` was >4100 lines long, a "Blob" module holding both core cold storage logic and over 2000 lines of tests. It made navigation and understanding the core operations difficult.
📐 Blueprint: Refactored into a `src/storage/redb_cold_storage/` module. Kept the core data definitions and implementation in `mod.rs` (reduced to ~2000 lines) and moved all tests to `tests.rs` (~2000 lines), maintaining the `#[cfg(test)]` block functionality but with better physical separation.
🧱 Stability: Reduced file sizes significantly, better modular separation for testing.
🔬 Verification: Builds successfully, all unit tests pass, and strict separation enforced.
